@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.6 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,10 +39,11 @@ dependencies:
   responsive_builder: ^0.7.0
   animated_text_kit: ^4.2.2
   carousel_slider: ^4.2.1
-  cupertino_icons: ^1.0.2
-  url_launcher: ^6.1.12
-  flutter_svg: ^2.0.7
-  get: ^4.6.5
+  cupertino_icons: ^1.0.6
+  url_launcher: ^6.2.2
+  flutter_svg: ^2.0.9
+  http: ^1.1.2
+  get: ^4.6.6
 
 dev_dependencies:
   flutter_test:
@@ -53,7 +54,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -65,23 +66,6 @@ flutter:
    assets:
      - assets/images/android.png
      - assets/images/jetpack_compose.png
-     - assets/images/dicoding_android_pemula.jpg
-     - assets/images/dicoding_kotlin.jpg
-     - assets/images/bookman.png
-     - assets/images/calculator.png
-     - assets/images/csnake.png
-     - assets/images/dujer.png
-     - assets/images/md2compose.png
-     - assets/images/md3compose.png
-     - assets/images/md3android.png
-     - assets/images/npuzzle.png
-     - assets/images/projman.png
-     - assets/images/remindme.png
-     - assets/images/saku.png
-     - assets/images/staver.png
-     - assets/images/tictactoe.png
-     - assets/images/musicompose.png
-     - assets/images/mathq.png
      - assets/images/skill_item_android.png
      - assets/images/skill_item_kotlin.png
      - assets/images/skill_item_jc.png
